Revert "style: global ai xiaona (#250)"
This reverts commit 594d8a14
