+++ src/views/home/index.vue
@@ -158,6 +158,12 @@
         </div>
       </div>
     </div>
+    <img
+      v-if="isMobile"
+      src="@/assets/img/xiaona.png"
+      class="absolute w-11 h-11 object-cover right-[2px] bottom-20 rounded-full shadow-lg arrow-box"
+      @click="onContactXN"
+    />
     <div
       v-if="showFooterContactVisible"
       class="home-contact-btn text-base absolute flex flex-col gap-[14px] overflow-hidden py-5 px-2 md:p-0 right-7 md:right-2"
@@ -276,6 +282,10 @@ const onLinkNash = () => {
     path: `/nash`
   })
   window.open(routerData.href, '_blank')
+}
+
+const onContactXN = () => {
+  window.open(`https://admin.sdlian.cn/l/ZkjRKDiEcb`, '_blank')
 }
 
 const videoMask = ref(false)
@@ -532,5 +542,22 @@ watch(
 .button-container {
   display: flex;
   gap: 1rem; /* Adjust the gap size as needed */
+}
+
+.arrow-box {
+  border: 2px solid rgba(124, 92, 252);
+  animation: blink 2s linear infinite;
+}
+
+@keyframes blink {
+  0% {
+    border-color: transparent;
+  }
+  50% {
+    border-color: rgba(124, 92, 252);
+  }
+  100% {
+    border-color: transparent;
+  }
 }
 </style>
